fix: Display of item type as 'link'
Migration to HNAPI related. Regular newsfeeds posts were of type 'story', but with the new API they're of type 'link'. This was causing all posts to have the word "link" after the comments count, when normally it'd only show the type if it was a job/ask post. Api docs here:
https://github.com/tastejs/hacker-news-pwas/blob/1a01152121c06e285445bc693f0d87a5fa555c7b/docs/api.md